+++ src/lib.rs
@@ -90,6 +90,7 @@ mod macros;
 pub mod asm;
 #[cfg(armv8m)]
 pub mod cmse;
+mod critical_section;
 pub mod delay;
 pub mod interrupt;
 #[cfg(all(not(armv6m), not(armv8m_base)))]